refactor(tui): break monolithic tui.go into 6 focused files
